+++ src/main/java/qna/domain/BaseTimeEntity.java
@@ -1,4 +1,4 @@
-package qnamission.domain;
+package qna.domain;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
