--- src/main/java/qnamission/domain/Answer.java
@@ -1,43 +1,0 @@
-package qnamission.domain;
-
-
-import javax.persistence.*;
-
-@Entity
-public class Answer extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Lob
-    private String contents;
-
-    @Column(nullable = false)
-    private boolean deleted = false;
-
-    private Long questionId;
-
-    private Long writerId;
-
-    protected Answer() {
-
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getContents() {
-        return contents;
-    }
-
-    public void setContents(String contents) {
-        this.contents = contents;
-    }
-
-}
