--- src/main/java/qnamission/domain/AnswerRepository.java
@@ -1,7 +1,0 @@
-package qnamission.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
-
-}
