--- src/main/java/qnamission/domain/DeleteHistoryRepository.java
@@ -1,6 +1,0 @@
-package qnamission.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface DeleteHistoryRepository extends JpaRepository<DeleteHistory, Long> {
-}
