--- src/main/java/qnamission/domain/QuestionRepository.java
@@ -1,6 +1,0 @@
-package qnamission.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface QuestionRepository extends JpaRepository<Question, Long> {
-}
