--- src/main/java/qnamission/domain/UserRepository.java
@@ -1,6 +1,0 @@
-package qnamission.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
