+++ src/main/resources/application.properties
@@ -1,11 +1,15 @@
-# spring.datasource.url=jdbc:h2:~/test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.url=jdbc:h2:~/test;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
 spring.datasource.username=sa
 spring.h2.console.enabled=true
 
 ## JPA
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=create
 
 # Mysql Dialect
-spring.datasource.url=jdbc:h2:~/test;MODE=MySQL
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL57Dialect
+#spring.datasource.url=jdbc:h2:~/test;MODE=MySQL
+#spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL57Dialect
+
+spring.output.ansi.enabled=always
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
