+++ src/test/java/qna/domain/AnswerRepositoryTest.java
@@ -1,21 +1,24 @@
-package qnamission.domain;
+package qna.domain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+
 import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
-class DeleteHistoryRepositoryTest {
+class AnswerRepositoryTest {
 
     @Autowired
-    private DeleteHistoryRepository deleteHistoryRepository;
+    private AnswerRepository answerRepository;
 
     @Test
     void save() {
-        final DeleteHistory expected = new DeleteHistory();
-        final DeleteHistory actual = deleteHistoryRepository.save(expected);
+        final Answer expected = new Answer();
+        expected.setContents("hello");
+        final Answer actual = answerRepository.save(expected);
         assertNotNull(actual.getId());
     }
+
 }
