+++ src/test/java/qna/domain/DeleteHistoryRepositoryTest.java
@@ -1,4 +1,4 @@
-package qnamission.domain;
+package qna.domain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,16 +7,16 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
-class QuestionRepositoryTest {
+class DeleteHistoryRepositoryTest {
 
     @Autowired
-    private QuestionRepository questionRepository;
+    private DeleteHistoryRepository deleteHistoryRepository;
 
     @Test
     void save() {
-        final Question expected = new Question();
-        expected.setTitle("test");
-        final Question actual = questionRepository.save(expected);
+        final DeleteHistory expected = new DeleteHistory();
+        final DeleteHistory actual = deleteHistoryRepository.save(expected);
         assertNotNull(actual.getId());
     }
+
 }
