+++ src/test/java/qna/domain/QuestionRepositoryTest.java
@@ -1,4 +1,4 @@
-package qnamission.domain;
+package qna.domain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,16 +7,17 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import static org.junit.jupiter.api.Assertions.*;
 
 @DataJpaTest
-class AnswerRepositoryTest {
+class QuestionRepositoryTest {
 
     @Autowired
-    private AnswerRepository answerRepository;
+    private QuestionRepository questionRepository;
 
     @Test
     void save() {
-        final Answer expected = new Answer();
-        expected.setContents("hello");
-        final Answer actual = answerRepository.save(expected);
+        final Question expected = new Question();
+        expected.setTitle("test");
+        final Question actual = questionRepository.save(expected);
         assertNotNull(actual.getId());
     }
+
 }
