+++ src/test/java/qna/domain/UserRepositoryTest.java
@@ -1,4 +1,4 @@
-package qnamission.domain;
+package qna.domain;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,17 +14,18 @@ class UserRepositoryTest {
 
     @Test
     void save() {
-        User expected = new User("name", "password", "userId");
+        User expected = new User("userId", "password", "name", "email");
         User actual = userRepository.save(expected);
         assertNotNull(actual.getId());
     }
 
     @Test
     void saveDuplicatedId() {
-        User user1 = new User("name", "password", "userId");
+        User user1 = new User("userId", "password1", "name1", "email1");
         userRepository.save(user1);
 
-        User user2 = new User("name1", "password1", "userId");
+        User user2 = new User("userId", "password2", "name2", "email2");
         assertThrows(Exception.class, () -> userRepository.save(user2));
     }
+
 }
